refactor: Share code between native and Skia activity
