+++ src/Uno.UI/UI/Xaml/ApplicationActivity.Android.cs
@@ -49,7 +49,6 @@ namespace Microsoft.UI.Xaml
 		private InputPane _inputPane;
 		private View _content;
 		private AWindow _window;
-		private OnBackPressedCallback _backPressedCallback;
 
 		public ApplicationActivity(IntPtr ptr, JniHandleOwnership owner) : base(ptr, owner)
 		{
@@ -281,44 +280,6 @@ namespace Microsoft.UI.Xaml
 			InitializeBackPressedCallback();
 		}
 
-		private void InitializeBackPressedCallback()
-		{
-			// On Android 15+ (API 35+), use OnBackPressedCallback for predictive back gesture support.
-			// The callback is enabled/disabled based on BackRequested subscription state.
-			if ((int)Build.VERSION.SdkInt >= 35)
-			{
-				_backPressedCallback = new UnoOnBackPressedCallback(
-					enabled: SystemNavigationManager.GetForCurrentView().HasBackRequestedSubscribers);
-				OnBackPressedDispatcher.AddCallback(this, _backPressedCallback);
-				SystemNavigationManager.BackRequestedSubscribersChanged += OnBackRequestedSubscribersChanged;
-			}
-		}
-
-		private void OnBackRequestedSubscribersChanged(object sender, bool hasSubscribers)
-		{
-			if (_backPressedCallback is not null)
-			{
-				_backPressedCallback.Enabled = hasSubscribers;
-			}
-		}
-
-		/// <summary>
-		/// Callback for handling back button presses on Android 15+ with predictive back gesture support.
-		/// </summary>
-		private sealed class UnoOnBackPressedCallback : OnBackPressedCallback
-		{
-			public UnoOnBackPressedCallback(bool enabled) : base(enabled)
-			{
-			}
-
-			public override void HandleOnBackPressed()
-			{
-				// On Android 15+, subscription to BackRequested means the app handles back navigation.
-				// The Handled property is ignored - back press is always consumed when callback is enabled.
-				SystemNavigationManager.GetForCurrentView().RequestBack();
-			}
-		}
-
 		private void OnInsetsChanged(Thickness insets)
 		{
 			NativeWindowWrapper.Instance.RaiseNativeSizeChanged();
@@ -382,12 +343,7 @@ namespace Microsoft.UI.Xaml
 			LayoutProvider.KeyboardChanged -= OnKeyboardChanged;
 			LayoutProvider.InsetsChanged -= OnInsetsChanged;
 
-			if (_backPressedCallback is not null)
-			{
-				SystemNavigationManager.BackRequestedSubscribersChanged -= OnBackRequestedSubscribersChanged;
-				_backPressedCallback.Remove();
-				_backPressedCallback = null;
-			}
+			CleanupBackPressedCallback();
 
 			NativeWindowWrapper.Instance.OnNativeClosed();
 		}
